Refactor file structure for the reports page
