+++ tests/e2e/reports-scenario.js
@@ -5,7 +5,7 @@
 //GetAttribute() returns "boolean" values and will return either "true" or null
 describe('Reports', function(){
     
-    var Reports = require('./pages/reports');
+    var Reports = require('../../app/reports/page.js');
     var reports = new Reports();
     
     describe('Reports List', function() {
